requirements.txt
aitools==0.1
playwright==1.46.0
python-dotenv==1.0.1